Support running a bone script from command-line
If command-line arguments are passed to bone, the first argument is
taken to be a name of the module (`.bn` suffix is added if not given).
Then that module is loaded and the REPL is not started.

Also, when loading a module, if the first line of the file starts with
'#', that line is skipped, to allow for executable bone scripts with a
shebang line.